requirements.txt
alembic==1.14.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.31.0
face-recognition==1.3.0
Flask==2.3.2
Flask-Migrate==4.0.6
Flask-SQLAlchemy==3.0.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
numpy==2.1.3
Pillow==9.5.0
python-dotenv==1.0.0
qrcode==7.3.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==2.3.6
flask-wtf==1.0.0